// frontend/src/api/model/region.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */

/**
 * 电影地区实体
 */
export interface Region {
  /** 地区ID */
  id?: number;
  /** 地区名称 */
  name?: string;
  /** 英文名称 */
  nameEn?: string;
  /** 地区描述 */
  description?: string;
  /** 创建时间 */
  createTime?: string;
}